+++ backend/src/main/java/com/project/backend/domain/book/service/KaKaoService.java
@@ -1,0 +1,15 @@
+package com.project.backend.domain.book.service;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+
+/**
+ * -- 카카오 도서 작업을 처리하는 서비스클래스 --
+ *
+ * @author -- 정재익 --
+ * @since -- 2월 4일 --
+ */
+@RequiredArgsConstructor
+@Service
+public class KaKaoService {
+}
